widl: Fix the conditions under which array freeing code is output.
It doesn't matter whether or not the array is declared as a pointer or
is declared using array subscripts - the array is still allocated by the
unmarshalling function and so needs to be freed.

Fix a typo that caused varying arrays never to be freed.

Fix another typo with the check for conformant arrays being performed
twice, redundantly.