Revert "widl: Use separated buffer for result in simplify_path."
This reverts commit cd954f93225128f352c80d680096f880749eccb7.

That commit was nonsense; this function is supposed to clean up
a path name in-place in the source buffer, if possible. By using
a separate buffer as the target, the caller of the function won't
notice the effect of it. The rest of the function expects to read
the string from the target buffer itself, e.g. the first call
after xmalloc is sanitise_path() on the newly allocated,
uninitialized result buffer, resulting in use of uninitialized data.

Signed-off-by: Martin Storsjo <martin@martin.st>